+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Code improvements
 * Update API to 5.11.0, tools to 5.5.1 and other packages"
 * Fix task fill gradient and keyboard focus styles
+* Fix sorting tasks when "Group tasks" is enabled
 
 ## 3.0.12
 * Fix not showing collapse button for parent tasks